Bump symfony dep injection component to ^3.4
That makes a number of components to be also updated to 3.4.x
and requires php >7.0.8 (instead of 7.0.0).

Also updated psalm-baseline.xml to ignore 'node-version' potential
wrong assignment values.

Note I've not run --update-baseline because it comes with other
changes that I don't want to modify now (removing some options).
I leave that for the experts because I'm not sure how the baseline
is decided and which part is automatic or manual.

A step before #21