Bump to codechecker/codenisffer 3.0
Changes in commands are important because all the old CLI machinery
in phpcs 3.0 has been removed. There were some different solutions
for this:

- start using the bin/phpcs binary (exec) and forget.
- create a custom runner/reporter like we did in local_codechecker.
- use exclusively the new phpcs 3.0 classes to get it done.

While I was really tempted to re-use the local_codechecker custom
runner / reporter, finally I went the pure phpcs classes way to
avoid creating any dependency on local_codechecker (apart from the
standard).

Unit tests have been slightly improved to better assert that all the
expected elements in the output are there.

Finally, we have had to instruct phpcs where the PHPCompatibility
standard is because, when running from PHAR, the CodeSniffer.conf
file bundled is not read and it's expected to be in real filesystem.
This was causing codechecher not to work from PHAR, so we are now
configuring it from the commands (Checker and Fixer).

Closes #49