Move phpcs command from library mode to binary mode
Until now we were using the phpcs in "library mode", i.e.
like it was a library and using its API to run it within
PHP. That had a few advantages but some problems, having
to "cheat" here and there to get everything working.

More noticeably, the include-paths and other details
were conflicting badly with the, now isolated and installed
by composer moodle-cs standard, and the phpcs.xml files now
present in core.

So this commit just changes that "unsupported" library use
by a more natural use as binary, so we just run it from shell,
with proper CLI options and so on.

The only detail a little bit tricky is that now, in order to
continue supporting the max-warnings option in moodle-plugin-ci,
instead of getting the number of errors and warnings from the API,
we have to generate the summary report and extract that information
from there. Anyway, it seems to be stable enough. Covered with tests.