Update to current box version (phar packaging)
Current version (2.x) is really old and emitting some
warnings with php >= 8.1. So, just move to current 4.x version.

Simply change to its new incarnation (fork of the original) and
tidy up a few parameters.

Note that the phar still has a good number of limitations:
- Not all checks work (phpcs, behat...).
- Cannot post-install moodle-local_ci (vnu.jar)
- ...

But all those are pre-existing and this new version does
not change that.

Note that box requires PHP 8.1 to build the phar, but the
phar works perfectly with PHP 7.4 (our min requirement).

Fixes #235