Bump self-tests to 4.1 requirements
Trivial for GHA, just change of images to postgres:12, but...

Travis changed the PostgreSQL package for version 11 and up so, instead of
using the "postgres" user it now uses the "travis" one. And the port is
5433 instead of 5432. We use the existence of the PGVER environmental
variable to decide which defaults to use.

More yet, the connection via "localhost" (local net) now requires login and
password, it used to be trust/peer auth mode (not requiring password). If we want to
keep localhost (+ port) working, then we need to edit the  pg_hba.conf file
to trust/peer the local connections and then restart the database.

So, at the end, we are going to use socket connections (host = '')
that is perfectly ok for Travis (non dockered database). Only if they
haven't been configured another way manually (user, host, port).