Ensure the parallel command is working ok
Actual problem is that some of the jobs like savepoints or
phpdoc do install code within the moodle checkout, in order
to be able to run:
- savepoints installs and removes local/plugin/check_upgrade_savepoints.php
- phpdoc installs and remoces local/moodlecheck

And for them it's a hard requirement to be in place to do their job.

When running in parallel, the presence and (near-immediate) removal of
those files, cause other jobs to become confused (phpmd, phpunit...).

So here we are creating "groups" of parallel runs (element is the
processes array. And we are running them (the groups) in parallel, but
1 by 1, so a group doesn't start running until the previous one
has ended.

That way we first run the "problematic" commands (savepoints and
phpdoc) and, once they have finished and the interim files have
been deleted, we can start with the second group of processes.