Bump moodle-local_moodlecheck to v1.1.4
Also, update moodle-cs base version (though that
one is automatic on composer install/project).

And fix a few fixtures to pass with new phpcs
and node 18 stuff. Note this causes tests for 3.x
branches to fail with grunt, because the generated
map files are now different from the 4.x ones. Hence,
we are preventing that with conditional regular
expression for all 3.x branches.