Bump all travis files and docs to require Ubuntu 20.04
NodeJS 18 requires GLIBC 2.28, that only comes in 18.10
(non LTS) and 20.04 (focal) Ubuntu versions. Change all the
Travis references to point to that, because, by default,
it uses denial older version.

Also bump the "make validate" self tests and the Travis
dist files and docs to use php74 as minimum,,
because focal on Travis doesn't support php73 any more.

Finally, it seems that, for focal, they have changed again
the PostgreSQL port to be 5432 (it was 5433 for PostgreSQL
11 and 12). So let's specify it again. Ping, pong!