Rename the "codechecker" command to "phpcs"
Just to be better aligned with other commands. The original name
comes from the times that we needed local_codechecker (that used
to contain the moodle standard). But that's not the case any more.

Of course, we keep the old "codechecker" working as command
alias, so this change should be 100% BC compatible with current
configurations / uses.