Add --max-warnings option to the PHPDoc command
To be on pair with the PHPCs command, now we support
the same option (--max-warnings xx) within PHPDocs.

Now, by defaults, warnings will be allowed (previously
they were leading to exit with error). And if anybody wants
to be stricter, adding --max-warnings=0 will do the trick.
(or any other number if there are a number of warnings to allow).

Covered with (dummy) tests, because the execution of the PHPDoc
command requires a complete moodle checkout.

Fixes #186