Move mooodle-local_ci dependencies to normal
When moodle-plugin-ci was incepted, moodle-local_ci still did
not have its own composer dependencies, so it was just installed
raw (as a "package" repository).

Since then, moodle-local_ci has got its own composer stuff, so
we now can change to install its dependencies "normally".

So, here we are, simply, switching from a "package" repository
to a "vcs" one. Pointing to the v1.0.21 release that is the
first supporting installation as a dependency.

And then, removing the annoying @composer update that was executed
in the post-x steps of composer, not needed anymore.

For more information, this issue contains all the related events:

https://github.com/moodlehq/moodle-local_ci/issues/280