Enable phpcs and phpcbf checks to run from phar
We need the 2 CLI scripts and the CodeSniffer.conf file
to be added to the phar, so we can "include" the scripts
for execution and copy the conf file temporarily to
be together with the phar (that's the way CodeSniffer
expects it).

Note this will reduce a little bit code-coverage but in
a future PR I want to add some integration tests to cover
the execution from phar, hardly can be done from unit tests.