Fix .env support when running from phar archive
With this changes, the moodle-plugin-ci execution
will be able to find a .env file in the same directory
that the .phar file containing it.

That way install will be able to create that file and
the rest of commands will access to that information
via dotenv, avoiding to have to export any env variable
manually for phar-based executions.