Remove CI testing on Travis
While we still keep moodle-plugin-ci officially supported
with Travis, it just so slow (for our current paid plan)
that make any branch or PR to take ages to be tested (~1 hour).

So here we are reducing the number of branches being tested to:
- master (highest PHP version supported).
- latest stable (highest PHP version supported).
- oldest stable supported (lowest PHP version supported).

That will save some precious minutes (7-8 per run) making things
a little bit more agile, specially given the falling relevance
of Travis since a couple of years ago (compared with the raise
of GHA).