Add the missing coreapis.txt file to the PHAR archive
Note that normally that file is used as fallback when
working offline, but better have it added to the PHAR.

Note that, soon, we'll replace that file by another one
because we have already lib/apis.json in core and we
will switch to it (better than crawling the devdocs).

Also, reorganise a little bit the CHANGELOG.md file.