Add back the selfupdate command (interim mode)
The self update command dependencies were removed in the
past but the command was left around. As a result of that
any attempt of running:

php moodle-plugin-ci selfcomplete

Is failing with PHP error (because the underlying humbug/phar-updater
library was not available any more.

So, with this commit, we are re-introducing it, but with the
following observations:

1. We have moved from humbug/phar-updater (archived) to laravel-zero/phar-updater
   that is maintained.
2. We are stuck to version 1.*, because of we still supporting PHP 7.4
3. I've run it with PHP 8.2 and it doesn't break badly, just some PHP
   deprecated notices are displayed. In any case, far better than the
   previous status, not working at all.
4. Having it back to code will allow us to further progress:
   - Getting it tested with some new integration test.
   - Have time to look if there is any other fork with relaxed
     requirements (or make our own one).
   - Try other selfupdate alternatives. Certainly the one we are
     currently using is the most advanced ones, although for basic
     usage (no rollback, no keys verification...) there are a few,
     with consolidation/self-update surely being the more active
     (again knowing that it's way simpler).

And that's the story of this PR, basically just keep the selfupdate
command working (knowing that emits a few PHP notices with PHP 8.x)
but allowing us to test it and try other alternatives in the future.