Add tests to cover what is testable in SelfUpdate command
Note that only helper methods can be tested for that command, not
the execution of the update itself (it has to happen within a PHAR).
And it doesn't make much sense to start mocking lots of stuff for that.

Some integration tests @ CIS will be in charge of
covering the real execution instead. Coming soon.