Bump various dependencies
- moodle-cs 3.4.1
- moodle-local_moodlecheck 1.3.0
- moodle-local_ci 1.0.28
- php-compatibility e5cd2e24

It also comes with a good number of updates in
dev tools and other dependencies.

Note that we have had to stay sticky to vimeo/psalm 5.19.*
because 5.20 comes with some new RiskyTruthyFalsyComparison
problems that we need to analyse. This issue has been created
about that: https://github.com/moodlehq/moodle-plugin-ci/issues/279