Exit on error when something was wrong executing the command
When something abnormal happens (wrong configuration, missing stuff,
connectivity problems, etc., the phpcs execution can end with error.

But, if max-warnings is enabled (>=0), then the count of errors and
warnings (that is always 0 on error) gets precedence, hiding the
real exit status. Hence it passes, when really it was not executed.

Now we detect those abnormal situations and exit with error when
something in the execution goes wrong.

Covered with tests, checking all the possible max-warnings combinations.